+++ src/index.html
@@ -7,7 +7,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Titillium+Web:wght@600&display=swap" rel="stylesheet">
-    <script src="./index.js" defer></script>
+    <script src="./index.js" type="modules"></script>
     <title>LeaderBoard</title>
 </head>
 <body>
